% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppSimpleTensor-package.r
\name{getTensorList}
\alias{getTensorList}
\title{Returns the list of current available compiled C++ tensors}
\usage{
getTensorList()
}
\description{
Returns the list of current available compiled C++ tensors
}
\examples{
getTensorList()
}
\keyword{compile}
\keyword{cpp}
\keyword{tensor}
