\name{print.tensorFunction}
\alias{print.tensorFunction}
\title{Function to print content of a tensorFunction}
\usage{
  \method{print}{tensorFunction} (tensor)
}
\arguments{
  \item{tensor}{tensor to print}
}
\description{
  Function to print content of a tensorFunction
}
\keyword{compile}
\keyword{cpp}
\keyword{tensor}

