\docType{package}
\name{RcppSimpleTensor}
\alias{RcppSimpleTensor}
\alias{RcppSimpleTensor-package}
\title{C++ tensor expressions made easy and fast}
\description{
  The goal of this library is to help create cpp
  implementations for multidimensional operations. Use the
  main 2 functions RcppSimpleTensor and TI to get started.
  ...
}
\author{
  Thibaut Lamadon \email{thibaut.lamadon@gmail.com} ,
  Florian Oswald <florian.oswald@gmail.com>
}
\references{
  \url{https://github.com/tlamadon/RcppSimpleTensor}
}

