% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppSimpleTensor.R
\name{createInlineTensor}
\alias{createInlineTensor}
\title{This function directly evaluates the tensor expression
using the arrays available in the current scope}
\usage{
createInlineTensor()
}
\description{
This function directly evaluates the tensor expression
using the arrays available in the current scope
}
