\name{print.inlineTensor}
\alias{print.inlineTensor}
\title{Prints the content of the inline tensor. Basically all c++ compiled tensors used by TI.}
\usage{
  \method{print}{inlineTensor} (TI)
}
\arguments{
  \item{TI}{inline tensor to print}
}
\description{
  Prints the content of the inline tensor. Basically all
  c++ compiled tensors used by TI.
}

