\name{print.tensorFunction}
\alias{print.tensorFunction}
\title{Function to print content of a tensorFunction}
\usage{
  print.tensorFunction(tensor)
}
\arguments{
  \item{tensor}{tensor to print source code}
}
\description{
  Function to print content of a tensorFunction
}
\keyword{compile}
\keyword{cpp}
\keyword{tensor}

