% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppSimpleTensor.R
\docType{package}
\name{RcppSimpleTensor}
\alias{RcppSimpleTensor}
\alias{RcppSimpleTensor-package}
\title{C++ tensor expressions made easy and fast}
\description{
The goal of this library is to help create cpp implementations for multidimensional operations.
Use the main 2 functions RcppSimpleTensor and TI to get started.
...
}
\references{
\url{https://github.com/tlamadon/RcppSimpleTensor}
}
\author{
Thibaut Lamadon \email{thibaut.lamadon@gmail.com} , Florian Oswald <florian.oswald@gmail.com>
}
